requirements.txt
rembg
ultralytics

# Computer Vision
opencv-python>=4.8.0
mediapipe>=0.10.30

# Математика и обработка данных
numpy>=1.24.0
scipy>=1.11.0

# Визуализация
matplotlib>=3.7.0
pillow>=10.0.0

# Утилиты
dataclasses-json==0.6.3